Abstracted out common code between benches and tests into a crate
